DELETEメソッドに所有者確認を実装

+++ app/api/item/delete/[id]/route.js
@@ -3,11 +3,17 @@ import connectDB from "@/app/utils/database";
 import { ItemModel } from "@/app/utils/schemaModels";
 
 export async function DELETE(request, context) {
+    const reqBody = await request.json()
     try {
         await connectDB()
         const params = await context.params
-        await ItemModel.deleteOne({_id: params.id})
-        return NextResponse.json({message: "アイテム削除成功"})
+        const singleItem = await ItemModel.findById(params.id)
+        if (singleItem.email === reqBody.email) {
+            await ItemModel.deleteOne({_id: params.id})
+            return NextResponse.json({message: "アイテム削除成功"})
+        } else {
+            return NextResponse.json({message: "他の人が作成したアイテムです"})    
+        }
     }catch {
         return NextResponse.json({message: "アイテム削除失敗"})
     }
